+++ tests/os_chmod_test.py
@@ -21,6 +21,11 @@ import pytest
 
 import pylint_secure_coding_standard as pylint_scs
 
+try:
+    from pylint.testutils import MessageTest
+except ImportError:
+    from pylint.testutils import Message as MessageTest
+
 
 class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
     CHECKER_CLASS = pylint_scs.SecureCodingStandardChecker
@@ -125,7 +130,7 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
         print(code)
         node = astroid.extract_node(code)
         if enabled_platform and forbidden != 'S_IRGRP':
-            with self.assertAddsMessages(pylint.testutils.Message(msg_id='os-chmod-unsafe-permissions', node=node)):
+            with self.assertAddsMessages(MessageTest(msg_id='os-chmod-unsafe-permissions', node=node)):
                 self.checker.visit_call(node)
         else:
             with self.assertNoMessages():