+++ tests/pickle_load_test.py
@@ -19,6 +19,11 @@ import pytest
 
 import pylint_secure_coding_standard as pylint_scs
 
+try:
+    from pylint.testutils import MessageTest
+except ImportError:
+    from pylint.testutils import Message as MessageTest
+
 
 class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
     CHECKER_CLASS = pylint_scs.SecureCodingStandardChecker
@@ -47,7 +52,7 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
     )
     def test_pickle_load_not_ok(self, s):
         node = astroid.extract_node(s + ' #@')
-        with self.assertAddsMessages(pylint.testutils.Message(msg_id='avoid-pickle-load', node=node)):
+        with self.assertAddsMessages(MessageTest(msg_id='avoid-pickle-load', node=node)):
             self.checker.visit_call(node)
 
     @pytest.mark.parametrize(
@@ -60,5 +65,5 @@ class TestSecureCodingStandardChecker(pylint.testutils.CheckerTestCase):
     )
     def test_pickle_open_importfrom(self, s):
         node = astroid.extract_node(s + ' #@')
-        with self.assertAddsMessages(pylint.testutils.Message(msg_id='avoid-pickle-load', node=node)):
+        with self.assertAddsMessages(MessageTest(msg_id='avoid-pickle-load', node=node)):
             self.checker.visit_importfrom(node)